Start separating HED 2-specific code
